refactor(rust): move the lmdb storage
to be able to compile rust doc examples to build a node
