+++ implementations/rust/ockam/ockam/src/lib.rs
@@ -57,6 +57,8 @@ pub mod workers;
 #[cfg(feature = "std")]
 pub use ockam_abac as abac;
 pub use ockam_identity as identity;
+#[cfg(feature = "std")]
+pub use ockam_identity::storage::lmdb_storage::*;
 
 pub use ockam_core::{
     allow, deny, errcode, route, Address, Any, AsyncTryClone, Encoded, Error, LocalMessage,
