+++ implementations/rust/ockam/ockam_abac/Cargo.toml
@@ -23,15 +23,20 @@ std = [
   "minicbor/std",
   "tracing/std",
   "either/use_std",
+  "lmdb",
   "once_cell/std",
   "regex",
+  "tokio",
   "wast",
 ]
+lmdb = ["tokio", "lmdb-rkv"]
 
 [dependencies]
 either = { version = "1.8.1", default-features = false }
+lmdb-rkv = { version = "0.14.0", optional = true }
 minicbor = { version = "0.19.0", features = ["derive", "alloc"] }
 ockam_core = { version = "0.78.0", path = "../ockam_core", default-features = false }
+ockam_executor = { version = "0.46.0", path = "../ockam_executor", default-features = false }
 ockam_identity = { version = "0.72.0", path = "../ockam_identity", default-features = false }
 once_cell = { version = "1.17.1", default-features = false, features = ["alloc"] }
 # optional:
@@ -39,6 +44,7 @@ regex = { version = "1.7.1", default-features = false, optional = true }
 rustyline = { version = "11.0.0", optional = true }
 rustyline-derive = { version = "0.8.0", optional = true }
 str-buf = "3.0.1"
+tokio = { version = "1.27", default-features = false, optional = true, features = ["sync", "time", "rt", "rt-multi-thread", "macros"] }
 tracing = { version = "0.1.34", default-features = false }
 wast = { version = "56.0.0", default-features = false, optional = true }
