+++ implementations/rust/ockam/ockam_abac/src/storage/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "std")]
+pub mod lmdb_storage;
+
+#[cfg(feature = "std")]
+pub use lmdb_storage::*;
