+++ implementations/rust/ockam/ockam_api/src/lib.rs
@@ -23,9 +23,6 @@ mod util;
 
 pub use util::*;
 
-#[cfg(feature = "lmdb")]
-pub mod lmdb;
-
 #[macro_use]
 extern crate tracing;
