+++ implementations/rust/ockam/ockam_identity/Cargo.toml
@@ -37,7 +37,10 @@ std = [
   "serde_bare/std",
   "minicbor/std",
   "time/std",
+  "lmdb",
 ]
+
+lmdb = ["tokio-retry", "lmdb-rkv"]
 
 # Feature: "no_std" enables functionality required for platforms
 # without the standard library.
@@ -66,6 +69,7 @@ cfg-if = "1.0.0"
 group = { version = "0.13.0", default-features = false }
 heapless = "0.7"
 hex = { version = "0.4", default-features = false }
+lmdb-rkv = { version = "0.14.0", optional = true }
 minicbor = { version = "0.19.0", features = ["alloc", "derive"] }
 ockam_core = { path = "../ockam_core", version = "^0.78.0", default-features = false }
 ockam_key_exchange_xx = { path = "../ockam_key_exchange_xx", version = "^0.74.0", default-features = false, optional = true }
@@ -80,6 +84,7 @@ serde_json = { version = "1.0", optional = true }
 sha2 = { version = "0.10", default-features = false }
 subtle = { version = "2.4.1", default-features = false }
 time = { version = "0.3.20", features = ["macros", "formatting", "std"], optional = true }
+tokio-retry = { version = "0.3.0", default-features = false, optional = true }
 tracing = { version = "0.1", default_features = false }
 
 [dev-dependencies]
