+++ implementations/rust/ockam/ockam_identity/src/identities/storage/lmdb_storage.rs
@@ -1,25 +1,27 @@
+use crate::Storage;
 use core::str;
 use lmdb::{Cursor, Database, Environment, Transaction};
-use minicbor::{Decode, Encode};
-use ockam::identity::identities::storage::Storage;
-use ockam_abac::{Action, Expr, PolicyStorage, Resource};
 use ockam_core::async_trait;
+use ockam_core::compat::boxed::Box;
+use ockam_core::compat::string::String;
+use ockam_core::compat::sync::Arc;
+use ockam_core::compat::vec::Vec;
 use ockam_core::errcode::{Kind, Origin};
 use ockam_core::{Error, Result};
 use ockam_node::tokio::task::{self, JoinError};
-use std::borrow::Cow;
 use std::fmt;
 use std::path::Path;
-use std::sync::Arc;
 use tokio_retry::strategy::{jitter, FixedInterval};
 use tokio_retry::Retry;
-use tracing as log;
+use tracing::debug;
 
-/// Lmdb AttributesStorage implementation
+/// Storage using the LMDB database
 #[derive(Clone)]
 pub struct LmdbStorage {
-    env: Arc<Environment>,
-    map: Database,
+    /// lmdb da
+    pub env: Arc<Environment>,
+    /// lmdb database file
+    pub map: Database,
 }
 
 impl fmt::Debug for LmdbStorage {
@@ -59,7 +61,8 @@ impl LmdbStorage {
         })
     }
 
-    async fn write(&self, k: String, v: Vec<u8>) -> Result<()> {
+    /// Write a new binary value for a given key in the database
+    pub async fn write(&self, k: String, v: Vec<u8>) -> Result<()> {
         let d = self.clone();
         let t = move || {
             let mut w = d.env.begin_rw_txn().map_err(map_lmdb_err)?;
@@ -71,7 +74,8 @@ impl LmdbStorage {
         task::spawn_blocking(t).await.map_err(map_join_err)?
     }
 
-    async fn delete(&self, k: String) -> Result<()> {
+    /// Delete a database entry
+    pub async fn delete(&self, k: String) -> Result<()> {
         let d = self.clone();
         let t = move || {
             let mut w = d.env.begin_rw_txn().map_err(map_lmdb_err)?;
@@ -129,80 +133,10 @@ impl Storage for LmdbStorage {
     }
 }
 
-/// Policy storage entry.
-///
-/// Used instead of storing plain `Expr` values to allow for additional
-/// metadata, versioning, etc.
-#[derive(Debug, Encode, Decode)]
-#[rustfmt::skip]
-struct PolicyEntry<'a> {
-    #[b(0)] expr: Cow<'a, Expr>
-}
-
-#[async_trait]
-impl PolicyStorage for LmdbStorage {
-    async fn get_policy(&self, r: &Resource, a: &Action) -> Result<Option<Expr>> {
-        let d = self.clone();
-        let k = format!("{r}:{a}");
-        let t = move || {
-            let r = d.env.begin_ro_txn().map_err(map_lmdb_err)?;
-            match r.get(d.map, &k) {
-                Ok(value) => {
-                    let e: PolicyEntry = minicbor::decode(value)?;
-                    Ok(Some(e.expr.into_owned()))
-                }
-                Err(lmdb::Error::NotFound) => Ok(None),
-                Err(e) => Err(map_lmdb_err(e)),
-            }
-        };
-        task::spawn_blocking(t).await.map_err(map_join_err)?
-    }
-
-    async fn set_policy(&self, r: &Resource, a: &Action, c: &Expr) -> Result<()> {
-        let v = minicbor::to_vec(PolicyEntry {
-            expr: Cow::Borrowed(c),
-        })?;
-        self.write(format!("{r}:{a}"), v).await
-    }
-
-    async fn del_policy(&self, r: &Resource, a: &Action) -> Result<()> {
-        self.delete(format!("{r}:{a}")).await
-    }
-
-    async fn policies(&self, r: &Resource) -> Result<Vec<(Action, Expr)>> {
-        let d = self.clone();
-        let r = r.clone();
-        let t = move || {
-            let tx = d.env.begin_ro_txn().map_err(map_lmdb_err)?;
-            let mut c = tx.open_ro_cursor(d.map).map_err(map_lmdb_err)?;
-            let mut xs = Vec::new();
-            for entry in c.iter_from(r.as_str()) {
-                let (k, v) = entry.map_err(map_lmdb_err)?;
-                let ks = str::from_utf8(k).map_err(from_utf8_err)?;
-                if let Some((prefix, a)) = ks.split_once(':') {
-                    if prefix != r.as_str() {
-                        break;
-                    }
-                    let x: PolicyEntry = minicbor::decode(v)?;
-                    xs.push((Action::new(a), x.expr.into_owned()))
-                } else {
-                    log::warn!(key = %ks, "malformed key in policy database")
-                }
-            }
-            Ok(xs)
-        };
-        task::spawn_blocking(t).await.map_err(map_join_err)?
-    }
-}
-
 fn map_join_err(err: JoinError) -> Error {
     Error::new(Origin::Application, Kind::Io, err)
 }
 
 fn map_lmdb_err(err: lmdb::Error) -> Error {
     Error::new(Origin::Application, Kind::Io, err)
-}
-
-fn from_utf8_err(err: str::Utf8Error) -> Error {
-    Error::new(Origin::Other, Kind::Invalid, err)
 }
