+++ implementations/rust/ockam/ockam_identity/src/identities/storage/mod.rs
@@ -1,8 +1,14 @@
 mod attributes_entry;
 mod identities_repository;
+/// LMDB implementation of the Storage trait
+#[cfg(feature = "std")]
+pub mod lmdb_storage;
 #[allow(clippy::module_inception)]
 mod storage;
 
 pub use attributes_entry::*;
 pub use identities_repository::*;
+
+#[cfg(feature = "std")]
+pub use lmdb_storage::*;
 pub use storage::*;
